Parse paste content with green-like highlighting
Delegates parsing to an struct to store whether an spoiler spans
multiple lines or not.

Makes use of the given epub-builder stylesheet to set a custom CSS file
to be used for most of the EPUB file. Said stylesheet is a part of
fimfic2epub's stylesheet used for story chapters.